Change staff unavailability implementation.
Currently this is implemented via setting extensions on the Person.
However, extensions are a bit unwieldy, and in particular this behaves
badly with UDFs. If you say

Define("Foo", ...)
SetStaffUnavailable([2005REYN01], Foo())

then that Define needs to be run again before you can call AssignStaff,
otherwise Foo() is an unknown name.